+++ tests/wps_restapi/test_processes.py
@@ -1,4 +1,5 @@
 # pylint: disable=R1729  # ignore non-generator representation employed for displaying test log results
+import uuid
 
 import base64
 import contextlib
@@ -45,10 +46,10 @@ from weaver.wps.utils import get_wps_url
 from weaver.wps_restapi import swagger_definitions as sd
 
 if TYPE_CHECKING:
-    from typing import Optional
+    from typing import Optional, Tuple
 
     from weaver.processes.constants import ProcessSchemaType
-    from weaver.typedefs import CWL, JSON
+    from weaver.typedefs import CWL, JSON, AnyHeadersContainer, AnyVersion
 
 
 # pylint: disable=C0103,invalid-name
@@ -617,11 +618,11 @@ class WpsRestApiProcessesTest(unittest.TestCase):
             assert proc["outputs"][0]["formats"][0]["mediaType"] == ContentType.APP_JSON
 
     def deploy_process_make_visible_and_fetch_deployed(self,
-                                                       deploy_payload,
-                                                       expected_process_id,
-                                                       headers=None,
-                                                       assert_io=True,
-                                                       ):
+                                                       deploy_payload,          # type: JSON
+                                                       expected_process_id,     # type: str
+                                                       headers=None,            # type: Optional[AnyHeadersContainer]
+                                                       assert_io=True,          # type: bool
+                                                       ):                       # type: (...) -> JSON
         """
         Deploy, make visible and obtain process description.
 
@@ -724,7 +725,12 @@ class WpsRestApiProcessesTest(unittest.TestCase):
         assert resp.status_code == 400
         assert "'Deploy.DeployCWL.id': 'Missing required field.'" in resp.json["cause"]
 
-    def deploy_process_CWL_direct(self, content_type, graph_count=0, process_id="test-direct-cwl-json", version=None):
+    def deploy_process_CWL_direct(self,
+                                  content_type,                         # type: ContentType
+                                  graph_count=0,                        # type: int
+                                  process_id="test-direct-cwl-json",    # type: str
+                                  version=None,                         # type: Optional[AnyVersion]
+                                  ):                                    # type: (...) -> Tuple[CWL, JSON]
         cwl_core = {
             "id": process_id,
             "class": "CommandLineTool",
@@ -775,7 +781,7 @@ class WpsRestApiProcessesTest(unittest.TestCase):
             "schema": {"type": "string", "contentMediaType": "text/plain"},
             "formats": [{"default": True, "mediaType": "text/plain"}]
         }]
-        return cwl
+        return cwl, desc
 
     def test_deploy_process_CWL_direct_JSON(self):
         self.deploy_process_CWL_direct(ContentType.APP_CWL_JSON)
@@ -1149,11 +1155,55 @@ class WpsRestApiProcessesTest(unittest.TestCase):
     def test_deploy_process_WPS3_DescribeProcess_executionUnit(self):
         raise NotImplementedError
 
+    def test_deploy_process_with_revision_invalid(self):
+        """
+        Ensure that new deployment directly using a ``{processID}:{version}`` reference is not allowed.
+
+        This nomenclature is reserved for revisions accomplished with PUT or PATCH requests with controlled versioning.
+
+        .. versionadded:: 4.20
+        """
+        cwl = {
+            "cwlVersion": "v1.0",
+            "class": "CommandLineTool",
+            "baseCommand": ["python3", "-V"],
+            "inputs": {},
+            "outputs": {
+                "output": {
+                    "type": "File",
+                    "outputBinding": {
+                        "glob": "stdout.log"
+                    },
+                }
+            },
+        }
+
+        headers = {"Content-Type": ContentType.APP_CWL_JSON, "Accept": ContentType.APP_JSON}
+        data = copy.deepcopy(cwl)
+        data["id"] = "invalid-process:1.2.3"
+        resp = self.app.post_json("/processes", params=cwl, headers=headers, expect_errors=True)
+        assert resp.status_code in [400, 422]
+        assert "invalid" in resp.json["description"]
+
+        data = {
+            "processDescription": {"process": {"id": "invalid-process:1.2.3"}},
+            "executionUnit": [{"unit": cwl}],
+            "deploymentProfileName": "http://www.opengis.net/profiles/eoc/dockerizedApplication",
+        }
+        resp = self.app.post_json("/processes", params=data, headers=self.json_headers, expect_errors=True)
+        assert resp.status_code in [400, 422]
+        assert "invalid" in resp.json["description"]
+
     def test_update_process_not_found(self):
         resp = self.app.patch_json("/processes/not-found", params={}, headers=self.json_headers, expect_errors=True)
         assert resp.status_code == 404
 
     def test_update_process_no_data(self):
+        """
+        Error expected if no data is provided for an update request.
+
+        .. versionadded:: 4.20
+        """
         p_id = "test-update-no-data"
         self.deploy_process_CWL_direct(ContentType.APP_JSON, process_id=p_id)
         resp = self.app.patch_json(f"/processes/{p_id}", params={}, headers=self.json_headers, expect_errors=True)
@@ -1169,27 +1219,39 @@ class WpsRestApiProcessesTest(unittest.TestCase):
         """
         Update the current process revision with new metadata (making it an older revision and new one becomes latest).
 
+        Change should be marked as PATCH revision.
+
         .. versionadded:: 4.20
         """
         p_id = "test-update-cwl-json"
-        self.deploy_process_CWL_direct(ContentType.APP_JSON, process_id=p_id)
+        _, desc = self.deploy_process_CWL_direct(ContentType.APP_JSON, process_id=p_id)
+        assert desc["process"]["version"] is None, "No version provided should be reported as such."
         data = {
             "description": "New description",
             "title": "Another title",
         }
         resp = self.app.patch_json(f"/processes/{p_id}", params=data, headers=self.json_headers)
         assert resp.status_code == 200
-        assert resp.json["processSummary"]["title"] == data["title"]
-        assert resp.json["processSummary"]["description"] == data["description"]
+        body = resp.json
+        assert body["processSummary"]["title"] == data["title"]
+        assert body["processSummary"]["description"] == data["description"]
 
         resp = self.app.get(f"/processes/{p_id}", headers=self.json_headers)
         assert resp.status_code == 200
-        assert resp.json["title"] == data["title"]
-        assert resp.json["description"] == data["description"]
+        body = resp.json
+        assert body["title"] == data["title"]
+        assert body["description"] == data["description"]
+        assert body["version"] == "0.0.1", (
+            "PATCH revision expected. Since previous did not have a version, "
+            "it should be assumed 0.0.0, making this revision 0.0.1."
+        )
 
     def test_update_process_older_valid(self):
         """
         Update an older process (already a previous revision) with new metadata.
+
+        The older and updated process references must then be adjusted to ensure that fetching by process name only
+        returns the new latest definition, while specific process tag returns the expected revision.
 
         .. versionadded:: 4.20
         """
@@ -1210,9 +1272,10 @@ class WpsRestApiProcessesTest(unittest.TestCase):
 
         resp = self.app.get(f"/processes/{p_id}", headers=self.json_headers)
         assert resp.status_code == 200
-        assert resp.json["version"] == "1.2.4", "Patch update expected"
-        assert resp.json["title"] == data["title"]
-        assert resp.json["description"] == data["description"]
+        body = resp.json
+        assert body["version"] == "1.2.4", "Patch update expected"
+        assert body["title"] == data["title"]
+        assert body["description"] == data["description"]
 
         data = {
             "title": "Another change with version",
@@ -1221,13 +1284,22 @@ class WpsRestApiProcessesTest(unittest.TestCase):
         resp = self.app.patch_json(f"/processes/{p_id}:{version}", params=data, headers=self.json_headers)
         assert resp.status_code == 200
 
+        # check that previous 'latest' can be fetched by specific version
+        resp = self.app.get(f"/processes/{p_id}:{version}", headers=self.json_headers)
+        assert resp.status_code == 200
+        body = resp.json
+        assert body["id"] == f"{p_id}:{version}"
+        assert body["version"] == version
+        assert "description" not in body
+
         # check final result with both explicit '1.2.7' version and new 'latest'
         for p_ref in [p_id, f"{p_id}:1.2.7"]:
             resp = self.app.get(f"/processes/{p_ref}", headers=self.json_headers)
             assert resp.status_code == 200
-            assert resp.json["version"] == data["version"], "Specific version update expected"
-            assert resp.json["title"] == data["title"]
-            assert "description" not in resp.json, (
+            body = resp.json
+            assert body["version"] == data["version"], "Specific version update expected"
+            assert body["title"] == data["title"]
+            assert "description" not in body, (
                 "Not modified since no new value, value from reference process must be used. "
                 "Must not make use of the intermediate '1.2.4' version, since '1.2.3' explicitly requested for update."
             )
@@ -1236,31 +1308,160 @@ class WpsRestApiProcessesTest(unittest.TestCase):
         """
         Validate that given a valid process update, associated jobs update their references to preserve links.
 
-        .. versionadded:: 4.20
-        """
-        raise NotImplementedError
-
-    def test_replace_process_latest_valid(self):
-        """
-        Redeploy a process by replacing its definition (MAJOR revision update).
+        If links were not updated with the new tagged revision, older jobs would refer to the updated (latest) process,
+        which might not make sense according to the level of modifications applied for this process.
 
         .. versionadded:: 4.20
         """
-        p_id = "test-update-cwl-json"
+        p_id = "test-update-job-refs"
         self.deploy_process_CWL_direct(ContentType.APP_JSON, process_id=p_id)
+        job = self.job_store.save_job(task_id=uuid.uuid4(), process=p_id)
+
+        # verify that job initially refers to "latest" process
+        path = f"/jobs/{job.id}"
+        resp = self.app.get(path, headers=self.json_headers)
+        body = resp.json
+        assert "processID" in body and body["processID"] == p_id
+        path = get_path_kvp(f"/processes/{p_id}/jobs", detail=False)
+        resp = self.app.get(path, headers=self.json_headers)
+        body = resp.json
+        assert len(body["jobs"]) > 1 and str(job.id) in body["jobs"]
+
+        # update process
         data = {
             "description": "New description",
             "title": "Another title",
         }
         resp = self.app.patch_json(f"/processes/{p_id}", params=data, headers=self.json_headers)
         assert resp.status_code == 200
-        assert resp.json["processSummary"]["title"] == data["title"]
-        assert resp.json["processSummary"]["description"] == data["description"]
+        p_rev = resp.json["version"]
+        assert p_rev is not None
+
+        # verify job was updated with new reference
+        path = f"/jobs/{job.id}"
+        resp = self.app.get(path, headers=self.json_headers)
+        body = resp.json
+        assert "processID" in body and body["processID"] == f"{p_id}:0.0.0"
+
+        path = get_path_kvp(f"/processes/{p_id}/jobs", detail=False)
+        resp = self.app.get(path, headers=self.json_headers)
+        body = resp.json
+        assert len(body["jobs"]) == 0
+        path = get_path_kvp(f"/processes/{p_id}:0.0.0/jobs", detail=False)
+        resp = self.app.get(path, headers=self.json_headers)
+        body = resp.json
+        assert len(body["jobs"]) == 1 and str(job.id) in body["jobs"]
+
+    def test_replace_process_valid(self):
+        """
+        Redeploy a process by replacing its definition (MAJOR revision update).
+
+        Validate both different deploy formats (CWL, OAS, OGC) and different resolution methods of target version based
+        auto-resolved latest process when omitted or using an explicit version specification in the payload.
+
+        .. versionadded:: 4.20
+        """
+        # first deploy uses direct CWL, following update uses OGC-AppPackage schema, and last used OpenAPI schema
+        # validate that distinct deployment schema does not pose a problem to parse update contents
+        p_id = "test-replace-cwl"
+        v1 = "1.2.3"
+        cwl_v1, desc_v1 = self.deploy_process_CWL_direct(ContentType.APP_JSON, process_id=p_id, version=v1)
+        assert desc_v1["process"]["version"] == v1
+        assert desc_v1["process"]["inputs"] == []
+
+        cwl_v2 = copy.deepcopy(cwl_v1)
+        cwl_v2.pop("id", None)  # ensure no reference
+        cwl_v2.pop("version", None)  # avoid conflict
+        cwl_v2["inputs"]["test"] = {"type": "string"}  # type: ignore
+        data = {
+            "processDescription": {"process": {
+                # must include ID in deploy payload for counter validation against updated process
+                # (error otherwise since reusing same Deploy schema that requires it)
+                "id": p_id,
+                # make public to allow retrieving the process
+                # since we "override" the revision with PUT, omitting this would make the new version private
+                "visibility": Visibility.PUBLIC,
+                # new information to apply, validate that it is also considered, not just the package redefinition
+                "title": "Updated CWL"
+            }},
+            "executionUnit": [{"unit": cwl_v2}],
+            "deploymentProfileName": "http://www.opengis.net/profiles/eoc/dockerizedApplication",
+        }
+        v2 = "2.0.0"  # not explicitly provided, expected resolved MAJOR update for revision
+        resp = self.app.put_json(f"/processes/{p_id}", params=data, headers=self.json_headers)
+        assert resp.status_code == 201
+        body = resp.json
+        assert body["processSummary"]["title"] == data["processDescription"]["process"]["title"], (
+            "Even though MAJOR update for CWL is accomplished, other fields that usually correspond to MINOR changes "
+            "should also applied at the same time since the operation replaces the new process definition (PUT)."
+        )
+        assert (
+            "description" not in body["processSummary"] or  # if undefined, dropped from body
+            body["processSummary"]["description"] != desc_v1["description"]  # just in case, check otherwise
+        ), (
+            "Description should not have remained from previous version since this is a replacement (PUT),"
+            "not a revision update (PATCH)."
+        )
+
+        path = get_path_kvp(f"/processes/{p_id}:{v1}", schema=ProcessSchema.OGC)
+        resp = self.app.get(path, headers=self.json_headers)
+        assert resp.status_code == 200
+        body = resp.json
+        assert "title" not in body, "should be missing as in original definition"
+        assert "description" not in body, "should be missing as in original definition"
+        assert body["version"] == v1
+        assert body["inputs"] == {}, "empty mapping due to OGC schema, no input as in original definition"
 
         resp = self.app.get(f"/processes/{p_id}", headers=self.json_headers)
         assert resp.status_code == 200
-        assert resp.json["title"] == data["title"]
-        assert resp.json["description"] == data["description"]
+        body = resp.json
+        assert body["title"] == data["processDescription"]["process"]["title"]
+        assert "description" not in body
+        assert body["version"] == v2, f"Since no version was specified, next MAJOR version after {v1} was expected."
+        assert len(body["inputs"]) == 1 and "test" in body["inputs"]
+
+        # redeploy with explicit version
+        cwl_v3 = copy.deepcopy(cwl_v2)
+        cwl_v3.pop("version", None)  # avoid conflict
+        # need to provide basic input definition in CWL to avoid dropping it when checked against payload definitions
+        # add extra literal data domain information in OAS structure
+        cwl_v3["inputs"]["number"] = {"type": "int"}  # type: ignore
+        v3 = "4.3.2"  # does not necessarily need to be the next one
+        data = {
+            "processDescription": {
+                "id": p_id,  # required to fulfill schema validation, must omit 'version' part and match request path ID
+                "version": "4.3.2",  # explicitly provided to avoid auto-bump to '3.0.0'
+                # use OAS representation in this case to validate it is still valid using update request
+                "inputs": {"number": {"schema": {"type": "integer", "minimum": 1, "maximum": 3}}},
+                "visible": Visibility.PUBLIC,  # ensure we can retrieve the description later
+            },
+            "executionUnit": [{"unit": cwl_v3}],
+        }
+        # don't need to refer to "latest" process since we provide an explicit version that is available
+        resp = self.app.put_json(f"/processes/{p_id}:{v1}", params=data, headers=self.json_headers)
+        assert resp.status_code == 201
+
+        # check all versions are properly resolved
+        resp = self.app.get(f"/processes/{p_id}:{v1}", headers=self.json_headers)
+        assert resp.status_code == 200
+        body = resp.json
+        assert body["version"] == v1
+        assert body["id"] == f"{p_id}:{v1}"
+        resp = self.app.get(f"/processes/{p_id}:{v2}", headers=self.json_headers)
+        assert resp.status_code == 200
+        body = resp.json
+        assert body["version"] == v2
+        assert body["id"] == f"{p_id}:{v2}"
+        resp = self.app.get(f"/processes/{p_id}:{v3}", headers=self.json_headers)  # explicitly the latest by version
+        assert resp.status_code == 200
+        body = resp.json
+        assert body["version"] == v3
+        assert body["id"] == p_id
+        resp = self.app.get(f"/processes/{p_id}", headers=self.json_headers)  # latest implicitly
+        assert resp.status_code == 200
+        body = resp.json
+        assert body["version"] == v3
+        assert body["id"] == p_id
 
     def test_delete_process_success(self):
         path = f"/processes/{self.process_public.identifier}"